Improve http state computation
Update the HTTP state through a HTTP policy and register it in the
azure operation pipeline.
Because we mainly want to have that on what has been needed and not on
what has been used on the network we register the policy on the operation
part and not the retry one (network issues will result in retry that can
increase the input/output).